requirements/dev.txt
-r common.txt
black==21.11b1
flake8==4.0.1
pep8-naming
flake8-broken-line
flake8-return
isort==5.10.1
django-debug-toolbar==3.2.3
pre-commit==2.16.0
